Prepend and postpent CRLF's to the GPS config.
The former will reset the GPS's decoder in case it has ingested some noisy bytes (note: many USB-to-serial converters spit out a bunch of garbage bytes when they are first turned on).
The latter will take care of the situation where the user has forgotten a trailing CRLF in their config (which would cause the last line in the config to not be observed).